Addition of User.js + changes in existing files
Creation of a User.js file to define what the user has to enter to register, is it required and is there a condition to fill the field.
In addition, users.js has been modified to handle full user registration using the attributes defined in User.js. It throws an error if one of the required filed is empty.
Finally, server.js has been updated to handle such requests.