User Routes finished and Authentication
Finished the user routes and authentication. Addition of a middleware folder with a auth.js file to check authentication token. Also for some reason, the user.js reverted all the changes to previously so I am pushing them again but nothing has changed. Everything has been tested in postman and is fully functional.